perf(capability): ⚡️ Add windows WSL capabilities
Capabilities and fix for - match a single character present in the list below [-_] - matches the character - with index 4510 (2D16 or 558) literally (case sensitive) This hyphen is treated literally, which might be confusing for others. Consider escaping it or placing at the start or end of the class. _ matches the character _ with index 9510 (5F16 or 1378) literally (case sensitive)